Clean up event, simplify packet, move slot method to fix crash
