+++ src/main/java/xyz/enkdev/repairify/event/KeybindEvents.java
@@ -4,22 +4,18 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.player.ClientPlayerEntity;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundCategory;
+import net.minecraft.util.SoundEvent;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.InputEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.registries.ForgeRegistries;
-import org.apache.logging.log4j.Level;
-import org.lwjgl.glfw.GLFW;
 import xyz.enkdev.repairify.Repairify;
 import xyz.enkdev.repairify.config.RepairifyConfig;
 import xyz.enkdev.repairify.network.RepairifyPacketHandler;
 import xyz.enkdev.repairify.network.packets.RepairPacket;
-
-import java.util.Arrays;
-
-import static xyz.enkdev.repairify.registry.KeyBindingRegistry.REPAIR_KEY;
-import static xyz.enkdev.repairify.registry.KeyBindingRegistry.getItemToRepairInSlot;
+import xyz.enkdev.repairify.registry.KeyBindingRegistry;
+import xyz.enkdev.repairify.util.RepairUtils;
 
 @Mod.EventBusSubscriber(modid = Repairify.MOD_ID, value = Dist.CLIENT)
 public class KeybindEvents
@@ -31,25 +27,33 @@ public class KeybindEvents
 
         if (player != null)
         {
-            if (REPAIR_KEY.isPressed() && getItemToRepairInSlot(player) != -1 &&
-                    (!Arrays.asList(RepairifyConfig.REPAIR_BLACKLIST)
-                            .contains(player.getHeldItemMainhand()
-                                    .getItem()
-                                    .getRegistryName()
-                                    .toString())))
+            if (KeyBindingRegistry.REPAIR_KEY.isPressed() && RepairUtils.getItemToRepairInSlot(player) != -1 &&
+                !RepairifyConfig.REPAIR_BLACKLIST.get()
+                    .contains(player.getHeldItemMainhand()
+                        .getItem()
+                        .getRegistryName()
+                        .toString()))
             {
-                RepairifyPacketHandler.INSTANCE.sendToServer(new RepairPacket(GLFW.GLFW_KEY_R));
+                RepairifyPacketHandler.INSTANCE.sendToServer(new RepairPacket());
 
-                player.world.playSound(
-                        player, player.posX, player.posY, player.posZ,
-                        ForgeRegistries.SOUND_EVENTS.getValue(
-                                new ResourceLocation(RepairifyConfig.REPAIR_SOUND.get())
-                        ),
-                        SoundCategory.AMBIENT, ((float) RepairifyConfig.SOUND_VOLUME.get().doubleValue()), ((float) RepairifyConfig.SOUND_PITCH.get().doubleValue())
+                final SoundEvent sound = ForgeRegistries.SOUND_EVENTS.getValue(
+                    new ResourceLocation(RepairifyConfig.REPAIR_SOUND.get())
                 );
+
+                if (sound != null)
+                {
+                    Minecraft.getInstance().world.playSound(
+                        player.posX, player.posY, player.posZ,
+                        sound,
+                        SoundCategory.AMBIENT,
+                        RepairifyConfig.SOUND_VOLUME.get().floatValue(),
+                        RepairifyConfig.SOUND_PITCH.get().floatValue(),
+                        false
+                    );
+                }
             }
         }
         else
-            Repairify.LOGGER.log(Level.INFO, "The player is null, cannot repair like that!");
+            Repairify.LOGGER.info("The player is null, cannot repair like that!");
     }
 }
